Core algoritm now works for individual buildings.
Intrestingly, there is a slight discrepancy between the output of the shortest path algorithm here and the same algorithm in the QGIS gui. Not sure why. The same stops appear to be calculated to be closer in QGIS. Since its a systematic and not random difference, I'm not concerned for now.