Rewrote routine for drawing cell borders.
Made the reference histogram for this routine the histogram
which counts all events in the full data set. This means you need
to calculate this histogram no matter what.

When sampling the data (e.g. picking wrist shots that were rebounds),
you could end up with a small sample size and then different "borders"
for the histogram, and I wanted to make all histograms have the same
borders.